move to shared/sessions
