--- kilo-sessions/README.md
@@ -1,3 +1,0 @@
-# @kilocode/sessions
-
-Session management package for KiloCode. Provides tools for managing session lifecycle including creation, syncing, restoration, and sharing.
