--- kilo-sessions/eslint.config.mjs
@@ -1,34 +1,0 @@
-import js from "@eslint/js"
-import eslintConfigPrettier from "eslint-config-prettier"
-import turboPlugin from "eslint-plugin-turbo"
-import tseslint from "typescript-eslint"
-
-export default [
-	js.configs.recommended,
-	eslintConfigPrettier,
-	...tseslint.configs.recommended,
-	{
-		plugins: {
-			turbo: turboPlugin,
-		},
-		rules: {
-			"turbo/no-undeclared-env-vars": "off",
-		},
-	},
-	{
-		rules: {
-			"@typescript-eslint/no-unused-vars": [
-				"error",
-				{
-					argsIgnorePattern: "^_",
-					varsIgnorePattern: "^_",
-					caughtErrorsIgnorePattern: "^_",
-				},
-			],
-			"@typescript-eslint/no-explicit-any": "error",
-		},
-	},
-	{
-		ignores: ["dist/*"],
-	},
-]
