--- kilo-sessions/tsconfig.json
@@ -1,34 +1,0 @@
-{
-	"compilerOptions": {
-		"target": "ES2022",
-		"module": "ESNext",
-		"moduleResolution": "bundler",
-		"allowSyntheticDefaultImports": true,
-		"esModuleInterop": true,
-		"declaration": true,
-		"declarationMap": true,
-		"sourceMap": true,
-		"outDir": "./dist",
-		"strict": true,
-		"noUnusedLocals": false,
-		"noUnusedParameters": false,
-		"exactOptionalPropertyTypes": true,
-		"noImplicitReturns": true,
-		"noFallthroughCasesInSwitch": true,
-		"noUncheckedIndexedAccess": true,
-		"noImplicitOverride": true,
-		"useUnknownInCatchVariables": true,
-		"forceConsistentCasingInFileNames": true,
-		"resolveJsonModule": true,
-		"isolatedModules": true,
-		"verbatimModuleSyntax": false,
-		"types": ["node"],
-		"baseUrl": ".",
-		"paths": {
-			"@roo/*": ["../src/shared/*"]
-		},
-		"skipLibCheck": true
-	},
-	"include": ["src"],
-	"exclude": ["node_modules", "dist", "**/*.test.ts", "**/*.spec.ts"]
-}
