+++ src/shared/kilocode/cli-sessions/core/SessionClient.ts
