+++ src/shared/kilocode/cli-sessions/core/TrpcClient.ts
