+++ src/shared/kilocode/cli-sessions/types/IApiConfig.ts
