+++ src/shared/kilocode/cli-sessions/types/IPathProvider.ts
